moved DetailsModal in client/components
