+++ src/client/components/DetailsModal/Details/DanishDetails.tsx
