+++ src/client/components/DetailsModal/Details/NorwegianDetails.tsx
