+++ src/client/components/DetailsModal/Details/SwedishApartmentDetails.tsx
