+++ src/client/components/DetailsModal/Details/SwedishHouseDetails.tsx
