+++ src/client/components/DetailsModal/Details/components/DetailInput.tsx
