+++ src/client/components/DetailsModal/Details/components/Details.styles.tsx
