+++ src/client/components/DetailsModal/Details/index.tsx
