+++ src/client/components/DetailsModal/index.tsx
