+++ src/client/components/DetailsModal/types.ts
