+++ src/client/pages/Checkout/CheckoutPayment/CheckoutPayment.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { useFormik, FormikHelpers } from 'formik'
 import { GraphQLError } from 'graphql'
 import { useTextKeys } from 'utils/textKeys'
-import { QuoteInput } from 'pages/Offer/Introduction/DetailsModal/types'
+import { QuoteInput } from 'components/DetailsModal/types'
 import { useCurrentLocale } from 'l10n/useCurrentLocale'
 import {
   useCreateQuoteBundleMutation,
