+++ src/client/pages/Checkout/CheckoutPayment/ContactInformation/ContactInformation.tsx
@@ -5,7 +5,7 @@ import { Headline } from 'components/Headline/Headline'
 import { useTextKeys } from 'utils/textKeys'
 import { useFeature, Features } from 'utils/hooks/useFeature'
 import { MEDIUM_SMALL_SCREEN_MEDIA_QUERY } from 'utils/mediaQueries'
-import { QuoteInput } from 'pages/Offer/Introduction/DetailsModal/types'
+import { QuoteInput } from 'components/DetailsModal/types'
 import { CreditCheckInfo } from 'pages/OfferNew/Checkout/CreditCheckInfo'
 import { TextInput, SsnInput } from 'pages/Offer/Checkout/inputFields'
 import { WrapperWidth } from '../../shared/CheckoutPageWrapper'
